requirements.txt
alembic==1.12.0
Babel==2.12.1
bcrypt==4.0.1
Beaker==1.12.1
certifi==2023.7.22
cffi==1.15.1
Chameleon==4.1.0
cryptography==41.0.3
dnspython==2.4.2
email-validator==2.0.0.post2
fanstatic==1.3
greenlet==2.0.2
hupper==1.12
idna==3.4
importlib-resources==6.0.1
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
PasteDeploy==3.0.1
Pillow==10.0.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
psycopg[c]==3.1.10
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
pyramid==2.0.2
pyramid-beaker==0.8
pyramid-chameleon==0.3
pyramid-layout==1.0
pyramid-retry==2.1.1
pyramid-tm==2.5
python-magic==0.4.27
pytz==2023.3.post1
sedate==1.0.3.post1
sentry-sdk==1.30.0
SQLAlchemy==2.0.20
transaction==3.1.0
translationstring==1.4
typing-extensions==4.7.1
urllib3==2.0.4
venusian==3.0.0
waitress==2.1.2
WebOb==1.8.7
WTForms==3.0.1
zope.deprecation==5.0
zope.event==5.0
zope.interface==6.0
zope.schema==7.0.1
zope.sqlalchemy==3.0

-e .